Time: 1 ms (97.94%) | Memory: 42.4 MB (88.79%) - LeetSync

+++ 69-sqrtx/sqrtx.java
@@ -2,14 +2,14 @@ class Solution {
     public int mySqrt(int x) {
         if(x < 2) return x;
 
-        int l = 0,r = x/2;
+        long l = 0,r = x/2;
         while(l <= r){
             long mid  = l+(r-l)/2;
             long square = mid*mid;
             if(square == x) return (int)mid;
-            else if(square < x) l = (int)mid+1;
-            else r = (int)mid-1; 
+            else if(square < x) l = mid+1;
+            else r = mid-1; 
         }
-        return r;
+        return (int)r;
     }
 }
